LayoutTests/inspector/profiler/canvas2d/canvas-replay-log-grid-expected.txt

Tests replay log grid.

Bug 109592 

Total frames: 2
Total draw call groups: 5
Total calls: 30
Total grid nodes: 37

Testing ReplayFirstStepClick
Testing ReplayNextStepClick
Testing ReplayPreviousStepClick
Testing ReplayNextDrawingCallClick starting on a frame node
Testing ReplayPreviousDrawingCallClick starting on a frame node
Testing ReplayNextDrawingCallClick starting on a draw call group
Testing ReplayPreviousDrawingCallClick starting on a call node
Testing ReplayNextDrawingCallClick starting on a call node

   |                   Frame #1 |                                | |
   |         Draw call group #1 |                                | |
 1 |                     save() | canvas-replay-log-grid.html:26 | |
 2 |                beginPath() | canvas-replay-log-grid.html:27 | |
 3 |       rect(0, 0, 100, 100) | canvas-replay-log-grid.html:28 | |
 4 |    context.fillStyle = red | canvas-replay-log-grid.html:29 | |
 5 |                     fill() | canvas-replay-log-grid.html:30 | |
   |         Draw call group #2 |                                | |
 6 |                  restore() | canvas-replay-log-grid.html:31 | |
 7 |                     save() | canvas-replay-log-grid.html:26 | |
 8 |                beginPath() | canvas-replay-log-grid.html:27 | |
 9 |         rect(5, 5, 95, 95) | canvas-replay-log-grid.html:28 | |
10 |  context.fillStyle = green | canvas-replay-log-grid.html:29 | |
11 |                     fill() | canvas-replay-log-grid.html:30 | |
12 |                  restore() | canvas-replay-log-grid.html:31 | |
   |                   Frame #2 |                                | |
   |         Draw call group #3 |                                | |
13 |                     save() | canvas-replay-log-grid.html:26 | |
14 |                beginPath() | canvas-replay-log-grid.html:27 | |
15 |       rect(10, 10, 90, 90) | canvas-replay-log-grid.html:28 | |
16 |   context.fillStyle = blue | canvas-replay-log-grid.html:29 | |
17 |                     fill() | canvas-replay-log-grid.html:30 | |
   |         Draw call group #4 |                                | |
18 |                  restore() | canvas-replay-log-grid.html:31 | |
19 |                     save() | canvas-replay-log-grid.html:26 | |
20 |                beginPath() | canvas-replay-log-grid.html:27 | |
21 |       rect(15, 15, 85, 85) | canvas-replay-log-grid.html:28 | |
22 | context.fillStyle = yellow | canvas-replay-log-grid.html:29 | |
23 |                     fill() | canvas-replay-log-grid.html:30 | |
   |         Draw call group #5 |                                | |
24 |                  restore() | canvas-replay-log-grid.html:31 | |
25 |                     save() | canvas-replay-log-grid.html:26 | |
26 |                beginPath() | canvas-replay-log-grid.html:27 | |
27 |       rect(20, 20, 80, 80) | canvas-replay-log-grid.html:28 | |
28 |  context.fillStyle = black | canvas-replay-log-grid.html:29 | |
29 |                     fill() | canvas-replay-log-grid.html:30 | |
30 |                  restore() | canvas-replay-log-grid.html:31 | |
